requirements.txt
Django
Pillow
psycopg2-binary
whitenoise
gunicorn
python-dotenv
python-decouple
dj-database-url
whitenoise
gunicorn